+++ app/src/main/kotlin/com/laralnet/agroai/account/domain/model/GoogleAccount.kt
@@ -1,8 +1,0 @@
-package com.laralnet.agroai.account.domain.model
-
-data class GoogleAccount(
-    val email: String,
-    val displayName: String,
-    val photoUri: String? = null,
-    val isSelected: Boolean = false
-)